[cmake] make velocypack an imported library
On systems with velocypack installed in a non-default location, the
include directory for velocypack as detected by the FindVelocypack.cmake
module was not added to the include path for the fuerte library target.

With this change the FindVelocypack.cmake module provides an imported
library target as recommended by modern CMake providing all required
information to link against it including include path.

Signed-off-by: Torbjörn Klatt <t.klatt.oss@mailbox.org>